Show a loading state and avoid stale content on log navigation
Clear pending sample summaries + the selected sample/details before paint when
the route's log changes (a shared hook used by both the sample-list and the
single-sample transcript routes), so a previous log's sample list or transcript
can't render under the next log's header while the new one loads. Show a loading
indicator on the eval page while a log loads instead of an empty panel.
