+++ apps/inspect/src/state/log.ts
@@ -1,4 +1,4 @@
-import { useCallback } from "react";
+import { useCallback, useLayoutEffect, useRef } from "react";
 
 import { useStore } from "./store";
 
@@ -10,11 +10,52 @@ export const useUnloadLog = () => {
     (state) => state.logsActions.clearSelectedLogFile
   );
   const clearLog = useStore((state) => state.logActions.clearLog);
+  const clearPendingSampleSummaries = useStore(
+    (state) => state.logActions.clearPendingSampleSummaries
+  );
 
   const unloadLog = useCallback(() => {
     clearSelectedLogDetails();
     clearSelectedLogFile();
     clearLog();
-  }, [clearLog, clearSelectedLogDetails, clearSelectedLogFile]);
+    clearPendingSampleSummaries();
+  }, [
+    clearLog,
+    clearSelectedLogDetails,
+    clearSelectedLogFile,
+    clearPendingSampleSummaries,
+  ]);
   return { unloadLog };
+};
+
+/**
+ * Clear log-bound state before paint when routing to another log. Pending
+ * summaries live outside selectedLogDetails but are merged into sample lists,
+ * so they must be reset with the selected sample/details.
+ */
+export const useClearStaleLogStateOnNav = (logPath: string | undefined) => {
+  const clearSelectedSample = useStore(
+    (state) => state.sampleActions.clearSelectedSample
+  );
+  const clearSelectedLogDetails = useStore(
+    (state) => state.logActions.clearSelectedLogDetails
+  );
+  const clearPendingSampleSummaries = useStore(
+    (state) => state.logActions.clearPendingSampleSummaries
+  );
+  const prevLogPathRef = useRef<string | undefined>(undefined);
+  useLayoutEffect(() => {
+    const prevLogPath = prevLogPathRef.current;
+    prevLogPathRef.current = logPath;
+    if (prevLogPath && logPath && logPath !== prevLogPath) {
+      clearSelectedSample();
+      clearSelectedLogDetails();
+      clearPendingSampleSummaries();
+    }
+  }, [
+    logPath,
+    clearSelectedSample,
+    clearSelectedLogDetails,
+    clearPendingSampleSummaries,
+  ]);
 };
